modules/loc-producer/requirements.txt
grpcio==1.40.0
kafka-python==2.0.2
protobuf==3.18.0
